# introduce canonical queries, use for normalization and dropck-outlives