--- a/src/librustc/lib.rs
+++ b/src/librustc/lib.rs
@@ -57,6 +57,7 @@
 #![feature(inclusive_range)]
 #![feature(inclusive_range_syntax)]
 #![cfg_attr(windows, feature(libc))]
+#![feature(macro_lifetime_matcher)]
 #![feature(macro_vis_matcher)]
 #![feature(match_default_bindings)]
 #![feature(never_type)]