… src/Cargo.lock
… src/librustc/infer/freshen.rs
… src/librustc/infer/mod.rs
… src/librustc/infer/outlives/obligations.rs
… src/librustc/infer/region_constraints/mod.rs
… src/librustc/lint/context.rs
… src/librustc/macros.rs
… src/librustc/mir/cache.rs
… src/librustc/mir/mod.rs
… src/librustc/mir/tcx.rs
… src/librustc/session/mod.rs
… src/librustc/traits/coherence.rs
… src/librustc/traits/error_reporting.rs
… src/librustc/traits/mod.rs
… src/librustc/traits/query/dropck_outlives.rs
… src/librustc/traits/query/mod.rs
… src/librustc/traits/query/normalize.rs
… src/librustc/traits/query/normalize_erasing_regions.rs
… src/librustc/traits/select.rs
… src/librustc/traits/specialize/mod.rs
… src/librustc/traits/structural_impls.rs
… src/librustc/traits/trans/mod.rs
… src/librustc/ty/context.rs
… src/librustc/ty/error.rs
… src/librustc/ty/flags.rs
… src/librustc/ty/fold.rs
… src/librustc/ty/instance.rs
… src/librustc/ty/layout.rs
… src/librustc/ty/maps/config.rs
… src/librustc/ty/maps/keys.rs