@nikomatsakis 2018-02-22
@bors try

I'd like to do a perf run.