@nikomatsakis 2018-02-22
cc @rust-lang/release -- I'd like to request a perf run. try build completed, not sure what other steps I need to do.