@nikomatsakis 2018-02-22
Here are the results:

[perf comparison](http://perf.rust-lang.org/compare.html?start=27a046e9338fb0455c33b13e8fe28da78212dedc&end=fc2b29dda6f30b073f45f1bd2fab8418ad8c68fd&stat=instructions%3Au)

Also archived here, since I've noticed that perf links tend to go stale:

http://smallcultfollowing.com/pr-48411/rustc%20performance%20data.htm