@nikomatsakis 2018-02-22
### Major changes (>10%)

Got worse:

- futures/futures-opt: +10-15%
- regex: 34.4% (patched incremental: compile one)
- regex-opt: 56% (patched incremental: compile one-opt)

Got better

- crates.io-opt: -30.0% on println-opt
- regression-31157, regression-31157-opt: -10% or so
- style-servo-opt: -21.0% (clean-incremental-opt)

### Medium changes (3-10%)

- coercions/coercions-opt: +4%
- piston-image: +6% (clean-incremental, clean-incremental-opt)
- unify-linearly: +9% (clean)

Got better:

- tokio-webpush-simple: -5%
- tokio-webpush-simple-opt: -2% through -5%
- unify-linearly: -4% (dummy fn)

### Small changes (<3%)

- syn: -2%
- crates.io: -2.5% mostly (one change +1%)
- deep-vector: +1%
- htm5ever: +2%
- html5ever-opt: +1%
- hyper: +1%, -2%
- parser, parser-opt: +2%
- piston-image: -1%, +6%
- tuple-stress: +2%