@nikomatsakis 2018-02-22
I'll try to take a look at `regex` and `regex-opt`.