@nikomatsakis 2018-02-24
New results available here: [instructions](http://perf.rust-lang.org/compare.html?start=063deba92e44809125a433ca6e6c1ad0993313bf&end=251f86541a2fcc1c26534ce4d680fe23f82a8bb7&stat=instructions%3Au), [wallclock](http://perf.rust-lang.org/compare.html?start=063deba92e44809125a433ca6e6c1ad0993313bf&end=251f86541a2fcc1c26534ce4d680fe23f82a8bb7&stat=wall-time)

Looking at those, and limiting to changes of >5%:

| Test | Improved? | Change | 
| -- | -- | -- |
| futures, futures-opt | 👿  | +11.9% |
| unify-linearly | 👿  | +7.2% |
| unify-linearly-opt | 👿 | 5.6% |
| unused-warnings | 👿 | 5.5% |
| unused-warnings-opt | 👿 | 5% |
| regression-31157-opt | 💚 | about -7% |
| tokio-webpush-simple | 💚 | about -6.6% |
| regression-31157 | 💚 | about -17% |