@nikomatsakis 2018-02-24
Interestingly, I can't reproduce the `futures` or `unify-linearly` measurements locally. I tried building with a plain `config.toml` and everything, but the runtimes just come out basically identical. I've not tried with the other tests yet.