@nikomatsakis 2018-02-25
@Mark-Simulacrum if it's easy to do, can you do another perf run, just so that we have up to date data?