@Mark-Simulacrum 2018-02-25
http://perf.rust-lang.org/compare.html?start=322d7f7b97f1672bade31e1c43d5753e0e45a1a8&end=94ed417a82daac50123a3e9ee3ed7cd2d308ba94&stat=instructions%3Au