@nikomatsakis 2018-02-27
cc @Mark-Simulacrum -- I'd like to see the impact of those latest changes. I found them to be a mild improvement locally, I want to see what the perf server thinks. (Sadly, the 3 or 4 other things I tried all seemed to be regressions.)