@nikomatsakis 2018-02-27
Note: the commits I pushed since the try are not important for perf, just new test cases...