@Mark-Simulacrum 2018-02-27
Hm, try build seems to have at least partially failed (https://travis-ci.org/rust-lang/rust/jobs/346967602#L9421-L9422), so @bors try