@nikomatsakis 2018-02-28
OK, to get tests to pass I had to rebase.

@bors try