@nikomatsakis 2018-02-28
cc @Mark-Simulacrum ok -- should be ready for perf comparison now