@nikomatsakis 2018-02-28
@Mark-Simulacrum will this be the URL?

http://perf.rust-lang.org/?start=29f5c699b11a6a148f097f82eaa05202f8799bbc&end=f1323669bf496891bbd821112eb5a6f89e15a283&absolute=true&stat=instructions%3Au