@nikomatsakis 2018-02-28
http://perf.rust-lang.org/compare.html?start=bedbad61195d2eae69b43eca49c6d3e2aee8f208&end=29f5c699b11a6a148f097f82eaa05202f8799bbc&stat=instructions:u