@nikomatsakis 2018-02-28
perf results available here : http://perf.rust-lang.org/compare.html?start=29f5c699b11a6a148f097f82eaa05202f8799bbc&end=f1323669bf496891bbd821112eb5a6f89e15a283&stat=instructions%3Au