@nikomatsakis 2018-02-28
My summary: 

This PR has a big effect on the futures test cases, for which it is an approximately ~20% hit. After a certain amount of profiling, I'm pretty sure that the problem here is that we are not able to get as much benefit from the existing projection cache as we used to get. That cache is tied to the inference context, and so when we move the "dropck-type-elaboration" into a query, it winds up running in another inference context, which doesn't have the cache. This hurts. I believe this will stop hurting when we change our caching scheme to be more global, as I plan to do.

It also regresses piston-image-check by about 10%. I have not investigated this case.

This can also be a substantial win, particularly in some incremental scenarios. For example, crates.io-opt is about 30% faster on the println case, and stylo-server-opt does much better on the "clean incremental" case. I imagine we are able to skip work we couldn't skip before. On the other hand, syn improves by about 7% across the board (and more with incremental) because we getting some caching we didn't get before (that was the primary benchmark I was looking at before).

Other tests are in the <5% range, and are a smatter of small improvements and small regressions.

I'm thinking that the best way forward may be to disable the dropck-elaboration-query **except** when NLL is enabled. I **believe** this would repair the regression for futures, while keeping the benefits for NLL. Moreover, it would let this canonicalization work land, which in turn unlocks progress on traits, which will hopefully let us recoup the perf hit across the board.